requirements.txt
alembic==1.4.2
attrs==19.3.0
blinker==1.4
Bootstrap-Flask==1.4
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
docopt==0.6.2
Email-Validator=1.1.1
Flask==1.1.2
Flask-CKEditor==0.4.4.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Share==0.1.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
more-itertools==8.4.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
